making multiselect way more readable by splitting it into multiple files
